# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Yuuka API",
    "description" : "Private paycheck-first budgeting API.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/templates/{templateId}/entries" : {
      "post" : {
        "tags" : [ "template-controller" ],
        "operationId" : "addEntry",
        "parameters" : [ {
          "name" : "templateId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateEntryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TemplateEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/{templateId}/entries/reorder" : {
      "post" : {
        "tags" : [ "template-controller" ],
        "operationId" : "reorder",
        "parameters" : [ {
          "name" : "templateId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReorderTemplateEntriesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TemplateResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/templates/{templateId}/archive" : {
      "post" : {
        "tags" : [ "template-controller" ],
        "operationId" : "archive",
        "parameters" : [ {
          "name" : "templateId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TemplateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paychecks" : {
      "post" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePaycheckRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaycheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paychecks/{paycheckId}/reopen" : {
      "post" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "reopen",
        "parameters" : [ {
          "name" : "paycheckId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaycheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paychecks/{paycheckId}/entries" : {
      "post" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "addEntry_1",
        "parameters" : [ {
          "name" : "paycheckId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateEntryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paychecks/{paycheckId}/entries/reorder" : {
      "post" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "reorder_1",
        "parameters" : [ {
          "name" : "paycheckId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReorderEntriesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaycheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paychecks/{paycheckId}/close" : {
      "post" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "close",
        "parameters" : [ {
          "name" : "paycheckId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaycheckResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/entries/{entryId}/status" : {
      "post" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "changeStatus",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StatusChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/template-entries/{entryId}" : {
      "delete" : {
        "tags" : [ "template-controller" ],
        "operationId" : "deleteEntry",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "version",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      },
      "patch" : {
        "tags" : [ "template-controller" ],
        "operationId" : "updateEntry",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TemplateEntryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TemplateEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paychecks/{paycheckId}" : {
      "get" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "paycheckId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "NOT_PAID", "PROCESSING", "POSTED" ]
          }
        }, {
          "name" : "type",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "BILL", "SPENDING_BUCKET", "SINKING_FUND" ]
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "custom"
          }
        }, {
          "name" : "ascending",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaycheckResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "archive_1",
        "parameters" : [ {
          "name" : "paycheckId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "version",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaycheckResponse"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "paycheckId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdatePaycheckRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaycheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entries/{entryId}" : {
      "delete" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "deleteEntry_1",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "version",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      },
      "patch" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "updateEntry_1",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateEntryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/paychecks/{paycheckId}/audit" : {
      "get" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "paycheckAudit",
        "parameters" : [ {
          "name" : "paycheckId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseAuditEventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paychecks/history" : {
      "get" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "history",
        "parameters" : [ {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "oldestFirst",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponsePaycheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/paychecks/active" : {
      "get" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "active",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponsePaycheckResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/entries/{entryId}/status-history" : {
      "get" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "statusHistory",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseStatusEventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entries/{entryId}/audit" : {
      "get" : {
        "tags" : [ "paycheck-controller" ],
        "operationId" : "entryAudit",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseAuditEventResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "TemplateEntryRequest" : {
        "type" : "object",
        "properties" : {
          "entryType" : {
            "type" : "string",
            "enum" : [ "BILL", "SPENDING_BUCKET", "SINKING_FUND" ]
          },
          "name" : {
            "type" : "string",
            "maxLength" : 160,
            "minLength" : 0
          },
          "defaultAmountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "defaultDueOffsetDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "accountName" : {
            "type" : "string",
            "maxLength" : 160,
            "minLength" : 0
          },
          "payee" : {
            "type" : "string",
            "maxLength" : 160,
            "minLength" : 0
          },
          "notes" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "targetMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "targetDate" : {
            "type" : "string",
            "format" : "date"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "entryType", "name" ]
      },
      "TemplateEntryResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entryType" : {
            "type" : "string",
            "enum" : [ "BILL", "SPENDING_BUCKET", "SINKING_FUND" ]
          },
          "name" : {
            "type" : "string"
          },
          "defaultAmountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          },
          "defaultDueOffsetDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "accountName" : {
            "type" : "string"
          },
          "payee" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "targetMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "targetDate" : {
            "type" : "string",
            "format" : "date"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "TemplateResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "archived" : {
            "type" : "boolean"
          },
          "entryCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "defaultTotalMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "entries" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TemplateEntryResponse"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "archivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "VersionRequest" : {
        "type" : "object",
        "properties" : {
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "version" ]
      },
      "ReorderTemplateEntriesRequest" : {
        "type" : "object",
        "properties" : {
          "entryIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "minItems" : 1
          },
          "templateVersion" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "entryIds", "templateVersion" ]
      },
...
      "CreatePaycheckRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "incomeDate" : {
            "type" : "string",
            "format" : "date"
          },
          "source" : {
            "type" : "string",
            "maxLength" : 160,
            "minLength" : 0
          },
          "notes" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          }
        },
        "required" : [ "amountMinor", "incomeDate", "name" ]
      },
...
      "CreateEntryRequest" : {
        "type" : "object",
        "properties" : {
          "entryType" : {
            "type" : "string",
            "enum" : [ "BILL", "SPENDING_BUCKET", "SINKING_FUND" ]
          },
          "name" : {
            "type" : "string",
            "maxLength" : 160,
            "minLength" : 0
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "accountName" : {
            "type" : "string",
            "maxLength" : 160,
            "minLength" : 0
          },
          "payee" : {
            "type" : "string",
            "maxLength" : 160,
            "minLength" : 0
          },
          "notes" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "targetMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "targetDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "amountMinor", "entryType", "name" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}